[cscore] Unify Cv and Raw source and sink as Image
