+++ cscore/src/main/java/edu/wpi/cscore/CameraServerCvJNI.java
@@ -59,11 +59,17 @@ public class CameraServerCvJNI {
     libraryLoaded = true;
   }
 
-  public static native int createCvSource(String name, int pixelFormat, int width, int height, int fps);
+  @Deprecated
+  public static int createCvSource(String name, int pixelFormat, int width, int height, int fps) {
+    return CameraServerJNI.createImageSource(name, pixelFormat, width, height, fps);
+  }
 
   public static native void putSourceFrame(int source, long imageNativeObj);
 
-  public static native int createCvSink(String name);
+  @Deprecated
+  public static int createCvSink(String name) {
+    return CameraServerJNI.createImageSink(name);
+  }
 
   public static native long grabSinkFrame(int sink, long imageNativeObj);
   public static native long grabSinkFrameTimeout(int sink, long imageNativeObj, double timeout);
