+++ cscore/src/main/java/edu/wpi/cscore/VideoSource.java
@@ -1,5 +1,5 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2016-2019 FIRST. All Rights Reserved.                        */
+/* Copyright (c) 2016-2020 FIRST. All Rights Reserved.                        */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -14,7 +14,7 @@ package edu.wpi.cscore;
  */
 public class VideoSource implements AutoCloseable {
   public enum Kind {
-    kUnknown(0), kUsb(1), kHttp(2), kCv(4), kRaw(8);
+    kUnknown(0), kUsb(1), kHttp(2), kImage(4), kCv(4), kRaw(4);
 
     @SuppressWarnings("MemberName")
     private final int value;
@@ -72,7 +72,7 @@ public class VideoSource implements AutoCloseable {
     switch (kind) {
       case 1: return Kind.kUsb;
       case 2: return Kind.kHttp;
-      case 4: return Kind.kCv;
+      case 4: return Kind.kImage;
       default: return Kind.kUnknown;
     }
   }
