+++ cscore/src/main/java/edu/wpi/cscore/raw/RawSink.java
@@ -1,5 +1,5 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2018-2019 FIRST. All Rights Reserved.                        */
+/* Copyright (c) 2018-2020 FIRST. All Rights Reserved.                        */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -25,7 +25,7 @@ public class RawSink extends ImageSink {
    * @param name Source name (arbitrary unique identifier)
    */
   public RawSink(String name) {
-    super(CameraServerJNI.createRawSink(name));
+    super(CameraServerJNI.createImageSink(name));
   }
 
   /**
