+++ cscore/src/main/native/include/cscore_c.h
@@ -17,6 +17,15 @@
 #endif
 
 #ifdef __cplusplus
+#include <wpi/deprecated.h>
+#else
+#ifndef WPI_DEPRECATED
+#define WPI_DEPRECATED(msg)  // nothing
+#endif
+
+#endif
+
+#ifdef __cplusplus
 extern "C" {
 #endif
 
@@ -126,8 +135,9 @@ enum CS_SourceKind {
   CS_SOURCE_UNKNOWN = 0,
   CS_SOURCE_USB = 1,
   CS_SOURCE_HTTP = 2,
-  CS_SOURCE_CV = 4,
-  CS_SOURCE_RAW = 8,
+  CS_SOURCE_IMAGE = 4,
+  CS_SOURCE_CV = CS_SOURCE_IMAGE,  // backwards compatibility
+  CS_SOURCE_RAW = CS_SOURCE_IMAGE  // backwards compatibility
 };
 
 /**
@@ -146,8 +156,9 @@ enum CS_HttpCameraKind {
 enum CS_SinkKind {
   CS_SINK_UNKNOWN = 0,
   CS_SINK_MJPEG = 2,
-  CS_SINK_CV = 4,
-  CS_SINK_RAW = 8
+  CS_SINK_IMAGE = 4,
+  CS_SINK_CV = CS_SINK_IMAGE,  // backwards compatibility
+  CS_SINK_RAW = CS_SINK_IMAGE  // backwards compatibility
 };
 
 /**
@@ -272,6 +283,9 @@ CS_Source CS_CreateHttpCamera(const char* name, const char* url,
 CS_Source CS_CreateHttpCameraMulti(const char* name, const char** urls,
                                    int count, enum CS_HttpCameraKind kind,
                                    CS_Status* status);
+CS_Source CS_CreateImageSource(const char* name, const CS_VideoMode* mode,
+                               CS_Status* status);
+WPI_DEPRECATED("use CS_CreateImageSource")
 CS_Source CS_CreateCvSource(const char* name, const CS_VideoMode* mode,
                             CS_Status* status);
 /** @} */
@@ -354,7 +368,7 @@ char** CS_GetHttpCameraUrls(CS_Source source, int* count, CS_Status* status);
 /** @} */
 
 /**
- * @defgroup cscore_opencv_source_cfunc OpenCV Source Functions
+ * @defgroup cscore_image_source_cfunc Image Source Functions
  * @{
  */
 void CS_NotifySourceError(CS_Source source, const char* msg, CS_Status* status);
@@ -377,6 +391,8 @@ void CS_SetSourceEnumPropertyChoices(CS_Source source, CS_Property property,
  */
 CS_Sink CS_CreateMjpegServer(const char* name, const char* listenAddress,
                              int port, CS_Status* status);
+CS_Sink CS_CreateImageSink(const char* name, CS_Status* status);
+WPI_DEPRECATED("use CS_CreateImageSink")
 CS_Sink CS_CreateCvSink(const char* name, CS_Status* status);
 /** @} */
 
@@ -411,7 +427,7 @@ int CS_GetMjpegServerPort(CS_Sink sink, CS_Status* status);
 /** @} */
 
 /**
- * @defgroup cscore_opencv_sink_cfunc OpenCV Sink Functions
+ * @defgroup cscore_image_sink_cfunc Image Sink Functions
  * @{
  */
 void CS_SetSinkDescription(CS_Sink sink, const char* description,
