+++ cscore/src/main/native/include/cscore_cpp.h
@@ -18,6 +18,7 @@
 #include <wpi/SmallVector.h>
 #include <wpi/StringRef.h>
 #include <wpi/Twine.h>
+#include <wpi/deprecated.h>
 
 #include "cscore_c.h"
 
@@ -205,6 +206,9 @@ CS_Source CreateHttpCamera(const wpi::Twine& name, const wpi::Twine& url,
 CS_Source CreateHttpCamera(const wpi::Twine& name,
                            wpi::ArrayRef<std::string> urls,
                            CS_HttpCameraKind kind, CS_Status* status);
+CS_Source CreateImageSource(const wpi::Twine& name, const VideoMode& mode,
+                            CS_Status* status);
+WPI_DEPRECATED("use CreateImageSource")
 CS_Source CreateCvSource(const wpi::Twine& name, const VideoMode& mode,
                          CS_Status* status);
 /** @} */
@@ -289,7 +293,7 @@ std::vector<std::string> GetHttpCameraUrls(CS_Source source, CS_Status* status);
 /** @} */
 
 /**
- * @defgroup cscore_opencv_source_func OpenCV Source Functions
+ * @defgroup cscore_image_source_func Image Source Functions
  * @{
  */
 void NotifySourceError(CS_Source source, const wpi::Twine& msg,
@@ -313,6 +317,8 @@ void SetSourceEnumPropertyChoices(CS_Source source, CS_Property property,
 CS_Sink CreateMjpegServer(const wpi::Twine& name,
                           const wpi::Twine& listenAddress, int port,
                           CS_Status* status);
+CS_Sink CreateImageSink(const wpi::Twine& name, CS_Status* status);
+WPI_DEPRECATED("use CreateImageSink")
 CS_Sink CreateCvSink(const wpi::Twine& name, CS_Status* status);
 /** @} */
 
@@ -353,7 +359,7 @@ int GetMjpegServerPort(CS_Sink sink, CS_Status* status);
 /** @} */
 
 /**
- * @defgroup cscore_opencv_sink_func OpenCV Sink Functions
+ * @defgroup cscore_image_sink_func Image Sink Functions
  * @{
  */
 void SetSinkDescription(CS_Sink sink, const wpi::Twine& description,
